refactor(util): extract getReferencePoint and asEdges into shared LineUtil
Move the reference-point helpers out of LabelBehavior into a shared
lib/util/LineUtil so they can be reused. No functional change.
