+++ lib/features/modeling/behavior/LabelBehavior.js
@@ -36,13 +36,9 @@ import {
 } from 'diagram-js/lib/util/PositionUtil';
 
 import {
-  sortBy
-} from 'min-dash';
-
-import {
-  getDistancePointLine,
-  perpendicularFoot
-} from './util/GeometricUtil';
+  getReferencePoint,
+  asEdges
+} from '../../../util/LineUtil';
 
 var NAME_PROPERTY = 'name';
 var TEXT_PROPERTY = 'text';
@@ -299,94 +295,8 @@ export function getReferencePointDelta(referencePoint, oldBounds, newBounds) {
 }
 
 /**
- * Generates the nearest point (reference point) for a given point
- * onto given set of lines
+ * Re-exported for backwards compatibility.
  *
- * @param {Point} point
- * @param {Line[]} lines
- *
- * @return {Point}
+ * @see module:lib/util/LineUtil
  */
-export function getReferencePoint(point, lines) {
-
-  if (!lines.length) {
-    return;
-  }
-
-  var nearestLine = getNearestLine(point, lines);
-
-  return perpendicularFoot(point, nearestLine);
-}
-
-/**
- * Convert the given bounds to a lines array containing all edges
- *
- * @param {Rect|Point} bounds
- *
- * @return {Line[]}
- */
-export function asEdges(bounds) {
-  return [
-    [ // top
-      {
-        x: bounds.x,
-        y: bounds.y
-      },
-      {
-        x: bounds.x + (bounds.width || 0),
-        y: bounds.y
-      }
-    ],
-    [ // right
-      {
-        x: bounds.x + (bounds.width || 0),
-        y: bounds.y
-      },
-      {
-        x: bounds.x + (bounds.width || 0),
-        y: bounds.y + (bounds.height || 0)
-      }
-    ],
-    [ // bottom
-      {
-        x: bounds.x,
-        y: bounds.y + (bounds.height || 0)
-      },
-      {
-        x: bounds.x + (bounds.width || 0),
-        y: bounds.y + (bounds.height || 0)
-      }
-    ],
-    [ // left
-      {
-        x: bounds.x,
-        y: bounds.y
-      },
-      {
-        x: bounds.x,
-        y: bounds.y + (bounds.height || 0)
-      }
-    ]
-  ];
-}
-
-/**
- * Returns the nearest line for a given point by distance
- * @param {Point} point
- * @param {Line[]} lines
- *
- * @return {Line}
- */
-function getNearestLine(point, lines) {
-
-  var distances = lines.map(function(l) {
-    return {
-      line: l,
-      distance: getDistancePointLine(point, l)
-    };
-  });
-
-  var sorted = sortBy(distances, 'distance');
-
-  return sorted[0].line;
-}
+export { getReferencePoint, asEdges } from '../../../util/LineUtil';
